Move bulk of the logic to the test runner
