+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/rewrite/testtool/LedgerApiTestTool.scala
@@ -3,60 +3,21 @@
 
 package com.daml.ledger.api.rewrite.testtool
 
-import java.util.concurrent.Executors
-
+import com.daml.ledger.api.rewrite.testtool.infrastructure.LedgerTestSuiteRunner
 import com.daml.ledger.api.rewrite.testtool.infrastructure.Reporter.ColorizedPrintStreamReporter
-import com.daml.ledger.api.rewrite.testtool.infrastructure.{
-  LedgerSession,
-  LedgerSessionConfiguration,
-  LedgerTestSuiteRunner
-}
-import com.daml.ledger.api.rewrite.testtool.tests.{Divulgence, Identity}
+import com.digitalasset.grpc.adapter.utils.DirectExecutionContext
 import org.slf4j.LoggerFactory
 
-import scala.concurrent.{ExecutionContext, Future}
-import scala.util.control.NonFatal
 import scala.util.{Failure, Success}
 
 object LedgerApiTestTool {
 
-  private val logger = LoggerFactory.getLogger(getClass.getName.stripSuffix("$"))
+  private[this] val logger = LoggerFactory.getLogger(getClass.getName.stripSuffix("$"))
 
   def main(args: Array[String]): Unit = {
+    val tests = new LedgerTestSuiteRunner(Cli.parse(args).getOrElse(sys.exit(1))).run()
 
-    val config = Cli.parse(args).getOrElse(sys.exit(1))
-
-    implicit val ec: ExecutionContext = {
-      val ec = ExecutionContext.fromExecutorService(Executors.newCachedThreadPool())
-      val _ = sys.addShutdownHook { val _ = ec.shutdownNow() }
-      ec
-    }
-
-    val configurations = Vector(
-      LedgerSessionConfiguration(config.host, config.port)
-    )
-
-    val sessions =
-      configurations.map(LedgerSession.getOrCreate).foldLeft(Vector.empty[LedgerSession]) {
-        (sessions, attempt) =>
-          attempt match {
-            case Failure(NonFatal(cause)) =>
-              logger.error(s"A ledger session could not be started, quitting...", cause)
-              sys.exit(1)
-            case Success(session) => sessions :+ session
-          }
-      }
-
-    val suiteConstructors = Vector(new Divulgence(_), new Identity(_))
-
-    val testSuites =
-      for (session <- sessions; suiteConstructor <- suiteConstructors)
-        yield suiteConstructor(session)
-    val runner = new LedgerTestSuiteRunner(ec)
-    val startedTests = Future.sequence(testSuites.flatMap(runner.run))
-
-    startedTests.onComplete(result => {
-      LedgerSession.closeAll()
+    tests.onComplete(result => {
       result match {
         case Success(results) =>
           new ColorizedPrintStreamReporter(System.out)(results)
@@ -65,8 +26,7 @@ object LedgerApiTestTool {
           logger.error("Unexpected uncaught exception, terminating!", e)
           sys.exit(1)
       }
-    })
-
+    })(DirectExecutionContext)
   }
 
 }
